review of protected renew application
