+++ frontend/app/routes/protected/renew/$id/confirm-home-address.tsx
@@ -245,7 +245,7 @@ export default function ProtectedRenewConfirmHomeAddress() {
           <Button variant="primary" id="continue-button" disabled={isSubmitting} data-gc-analytics-customclick="ESDC-EDSC:CDCP Renew Application Form-Adult:Save - Update address click">
             {t('protected-renew:update-address.save-btn')}
           </Button>
-          <ButtonLink id="back-button" routeId="public/renew/$id/ita/review-information" params={params} disabled={isSubmitting} data-gc-analytics-customclick="ESDC-EDSC:CDCP Renew Application Form-Adult:Cancel - Update address click">
+          <ButtonLink id="back-button" routeId="protected/renew/$id/review-adult-information" params={params} disabled={isSubmitting} data-gc-analytics-customclick="ESDC-EDSC:CDCP Renew Application Form-Adult:Cancel - Update address click">
             {t('protected-renew:update-address.cancel-btn')}
           </ButtonLink>
         </div>
