+++ frontend/public/locales/en/protected-renew.json
@@ -522,7 +522,6 @@
     "demographic-survey-change": "Change answers to demographic questions",
     "continue-button": "Continue",
     "back-button": "Back",
-    "submit-button": "Submit renewal application",
     "yes": "Yes",
     "no": "No"
   },
