+++ frontend/public/locales/fr/protected-renew.json
@@ -522,7 +522,6 @@
     "demographic-survey-change": "Change answers to demographic questions",
     "continue-button": "Continuer",
     "back-button": "Retour",
-    "submit-button": "Présenter la demande de renouvellement",
     "yes": "Oui",
     "no": "Non"
   },
